Hito3

+++ kernel/include/planificador_corto.h
@@ -10,7 +10,7 @@ void *planificador_corto_plazo_hrrn(void *_);
 void *esperar_salida_exec(void *_);
 void *esperar_salida_block(void *multiprocesamiento_p);
 void *hilo_liberar_multiprocesamiento(void *multiprocesamiento_p);
-int calcular_response_ratio(t_proceso *proceso);
+float calcular_response_ratio(t_proceso *proceso);
 void *esperar_bloqueo(void *multiprocesamiento_p);
 
 #endif
